Minor updates
- Converted float usage to doubles
- Converted world usages to ModifiableWorlds
- Added extra conversion methods to Quaternion

+++ src/main/java/com/cumulus/shapes/impl/Shapes.java
@@ -9,7 +9,7 @@ import com.cumulus.shapes.api.Shape;
  */
 public class Shapes {
 
-    public static Shape rectangle(float w, float d) {
+    public static Shape rectangle(double w, double d) {
         return Shape.of(
                 (pos) -> pos.getY() > 0 && pos.getY() <= 1,
                 Position.of(w/2, 0, d/2),
@@ -17,7 +17,7 @@ public class Shapes {
         );
     }
 
-    public static Shape ellipse(float a, float b) {
+    public static Shape ellipse(double a, double b) {
         return Shape.of(
                 (pos) -> ((pos.getX() * pos.getX())/(a * a)) + ((pos.getZ() * pos.getZ())/(b * b)) < 1 && pos.getY() > 0 && pos.getY() < 1,
                 Position.of(a, 0, b),
@@ -28,7 +28,7 @@ public class Shapes {
 
 
 
-    public static Shape ellipticalPrism(float a, float b, float height) {
+    public static Shape ellipticalPrism(double a, double b, double height) {
         return Shape.of(
                 (pos) -> ((pos.getX() * pos.getX())/(a * a)) + ((pos.getZ() * pos.getZ())/(b * b)) < 1 && pos.getY() > -height/2 && pos.getY() < height/2,
                 Position.of(a, height/2, b),
@@ -36,7 +36,7 @@ public class Shapes {
         );
     }
 
-    public static Shape rectanglarPrism(float width, float height, float depth) {
+    public static Shape rectanglarPrism(double width, double height, double depth) {
         return Shape.of(
                 (pos) -> pos.getX() > -width/2 && pos.getX() < width/2 && pos.getY() > -height/2 && pos.getY() < height/2 && pos.getZ() > -depth/2 && pos.getZ() < depth/2,
                 Position.of(width/2, height/2, depth/2),
@@ -44,7 +44,7 @@ public class Shapes {
         );
     }
 
-    public static Shape triangularPrism(float width, float height, float depth) {
+    public static Shape triangularPrism(double width, double height, double depth) {
         return Shape.of(
                 (pos) -> pos.getX() > -width/2 && pos.getX() < width/2 && pos.getY() > -height/2 && pos.getY() < height/2 && pos.getZ() > -(depth*(1-((pos.getY()+height/2)/height)))/2 && pos.getZ() < (depth*(1-((pos.getY()+height/2)/height)))/2,
                 Position.of(width/2, height/2, depth/2),
@@ -56,7 +56,7 @@ public class Shapes {
 
 
 
-    public static Shape rectangularPyramid(float width, float height, float depth) {
+    public static Shape rectangularPyramid(double width, double height, double depth) {
         return Shape.of(
                 (pos) -> pos.getX() > -(width*(1-(pos.getY()/height)))/2 && pos.getX() < (width*(1-(pos.getY()/height)))/2 && pos.getY() > 0 && pos.getY() < height && pos.getZ() > -(depth*(1-(pos.getY()/height)))/2 && pos.getZ() < (depth*(1-(pos.getY()/height)))/2,
                 Position.of(width/2, height, depth/2),
@@ -64,7 +64,7 @@ public class Shapes {
         );
     }
 
-    public static Shape ellipticalPyramid(float a, float b, float height) {
+    public static Shape ellipticalPyramid(double a, double b, double height) {
         return Shape.of(
                 (pos) -> ((pos.getX() * pos.getX())/((a*(1-(pos.getY()/height))) * (a*(1-(pos.getY()/height))))) + ((pos.getZ() * pos.getZ())/((b*(1-(pos.getY()/height))) * (b*(1-(pos.getY()/height))))) < 1 && pos.getY() > 0 && pos.getY() < height,
                 Position.of(a, height, b),
@@ -73,7 +73,7 @@ public class Shapes {
     }
 
 
-    public static Shape ellipsoid(float a, float b, float c) {
+    public static Shape ellipsoid(double a, double b, double c) {
         return Shape.of(
                 (pos) -> ((pos.getX() * pos.getX())/(a * a)) + ((pos.getZ() * pos.getZ())/(b * b)) + ((pos.getY() * pos.getY())/(c * c)) < 1,
                 Position.of(a, c, b),
@@ -81,7 +81,7 @@ public class Shapes {
         );
     }
 
-    public static Shape hemiEllipsoid(float a, float b, float c) {
+    public static Shape hemiEllipsoid(double a, double b, double c) {
         return Shape.of(
                 (pos) -> ((pos.getX() * pos.getX())/(a * a)) + ((pos.getZ() * pos.getZ())/(b * b)) + ((pos.getY() * pos.getY())/(c * c)) < 1 && pos.getY() > 0,
                 Position.of(a, c, b),
